Refactoring. Add NoleggioService.
