+++ src/main/java/it/unisa/c02/rently/rently_application/business/gestioneNoleggio/service/GestioneNoleggioService.java
@@ -20,5 +20,5 @@ public interface GestioneNoleggioService {
 
 
     Noleggio getNoleggio(long id);
-    boolean checkFineNoleggio(Noleggio noleggio);
+    List<Noleggio> checkFineNoleggio(Date dateNow);
 }
